making functions.lua and resizer listener
moving all functions to functions.lua and creating a resizer listener
